+++ player.h
@@ -5,8 +5,8 @@ using namespace Tmpl8;
 class Player : public GameObject {
 public:
 	// constructor / destructor
-	Player(Sprite* sprite, float2 position);
-	Player(Sprite* sprite, float2 position, float speed);
+	Player(Sprite* sprite, int2 position);
+	Player(Sprite* sprite, int2 position, float speed);
 	~Player();
 
 	void Update(Input* input, float dt);
